Fixed a major design error
We now no longer format script output. I've added the functionality that
this change cost back with the ability to set bard variables to
environment variables.

All in all this is much cleaner and should make is much easier to write
scripts, while keeping them completely separated from your colors and
fonts. This was something that i needed to do anyway.

This change also bring a huge change to the way formatting is done.
Previously the string was parsed (formatting or otherwise) every time we
wanted formatted. We now compile it once (basically splitting it into
parts that are static and parts that are dynamic) and stitch those
together when we know what the user wants. It should bring a small speed
boost to formatting.