Switched error reporting structure once again
Now it's using something custom. Let's hope I like it more this time. At
least it looks better and is less cumbersome. The errors it spits out
are also a lot more useful, and since they are calling internal
function I'm having an easier time setting breakpoints.

I need to do something to allow the program to eat error and carry on,
but that's for the future. Right now error are only for fatal
unrecoverable things